Move utility classes to separate package
