+++ src/utility/NoHitException.java
@@ -1,0 +1,8 @@
+package utility;
+
+public class NoHitException extends Exception {
+    public NoHitException(String message) {
+        super(message);
+    }
+}
+
