+++ src/utility/RaycastObject.java
@@ -1,4 +1,4 @@
-package algorithm;
+package utility;
 
 import javafx.geometry.Point2D;
